pyproject.toml
[project]
name = "ai-fitness-trainer"
version = "0.1.0"
description = "Personal AI fitness trainer Telegram bot powered by Claude"
requires-python = ">=3.14"
dependencies = [
    "aiogram>=3.25.0",
    "aiosqlite>=0.22.0",
    "python-dotenv>=1.2.0",
    "apscheduler>=3.11.0",
]
